Create field-kit-utils pkg.
