+++ packages/field-scripts/build-module/bundler.js
@@ -5,8 +5,8 @@ import createVuePlugin from '@vitejs/plugin-vue';
 import envCompatible from 'vite-plugin-env-compatible';
 import { viteCommonjs } from '@originjs/vite-plugin-commonjs';
 import jsYaml from 'js-yaml';
+import { snake } from 'field-kit-utils/string-case.js';
 import fieldModulePlugin from './rollup-plugin-field-module.js';
-import { snake } from './string-case.js';
 
 const writeYaml = (filepath, data) => {
   const yaml = jsYaml.dump(data);
