+++ packages/field-scripts/package.json
@@ -16,6 +16,7 @@
     "@originjs/vite-plugin-commonjs": "^1.0.3",
     "@vitejs/plugin-vue": "^2.2.2",
     "field-kit": "file:../..",
+    "field-kit-utils": "file:../field-kit-utils",
     "js-yaml": "^4.1.0",
     "ramda": "^0.28.0",
     "vite": "^2.8.4",
